requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
openai==1.51.0
gspread==6.1.2
google-auth==2.35.0
